+++ src/process.rs
@@ -20,6 +20,8 @@ pub(crate) struct ProcessBuilder {
     program: OsString,
     /// A list of arguments to pass to the program.
     args: Vec<OsString>,
+    /// The length of arguments that is not replaced by `args_replace`.
+    base_args: usize,
     /// The environment variables in the expression's environment.
     env: BTreeMap<String, Option<OsString>>,
     /// The working directory where the expression will execute.
@@ -34,6 +36,7 @@ impl ProcessBuilder {
         Self {
             program: program.into(),
             args: Vec::new(),
+            base_args: 0,
             env: BTreeMap::new(),
             dir: None,
             stdout_to_stderr: false,
@@ -41,8 +44,15 @@ impl ProcessBuilder {
     }
 
     /// (chainable) Adds `arg` to the args list.
-    pub(crate) fn arg(&mut self, arg: impl AsRef<OsStr>) -> &mut Self {
-        self.args.push(arg.as_ref().to_os_string());
+    pub(crate) fn base_arg(&mut self, arg: impl Into<OsString>) -> &mut Self {
+        self.args.insert(self.base_args, arg.into());
+        self.base_args += 1;
+        self
+    }
+
+    /// (chainable) Adds `arg` to the args list.
+    pub(crate) fn arg(&mut self, arg: impl Into<OsString>) -> &mut Self {
+        self.args.push(arg.into());
         self
     }
 
@@ -57,7 +67,8 @@ impl ProcessBuilder {
         &mut self,
         args: impl IntoIterator<Item = impl AsRef<OsStr>>,
     ) -> &mut Self {
-        self.args = args.into_iter().map(|t| t.as_ref().to_os_string()).collect();
+        self.args.truncate(self.base_args);
+        self.args.extend(args.into_iter().map(|t| t.as_ref().to_os_string()));
         self
     }